Moved TileSet from GameMap to TileContainer as requested by @oyvindln
