+++ source/gamemap/GameMap.h
@@ -28,7 +28,6 @@
 #endif //mode_t
 #endif //mingw32
 
-#include <OgreVector3.h>
 #include <map>
 #include <string>
 #include <cstdint>
@@ -49,70 +48,10 @@ class CreatureMood;
 class RenderedMovableEntity;
 class Room;
 class Spell;
-class TileSet;
 
 enum class GameEntityType;
 enum class FloodFillType;
 enum class RoomType;
-enum class TileVisual;
-
-class TileSetValue
-{
-public:
-    TileSetValue() :
-        mRotationX(0.0),
-        mRotationY(0.0),
-        mRotationZ(0.0)
-    {}
-
-    TileSetValue(const std::string& meshName, const std::string& materialName,
-            Ogre::Real rotationX, Ogre::Real rotationY,Ogre::Real rotationZ) :
-        mMeshName(meshName),
-        mMaterialName(materialName),
-        mRotationX(rotationX),
-        mRotationY(rotationY),
-        mRotationZ(rotationZ)
-    {}
-
-    inline const std::string& getMeshName() const
-    { return mMeshName; }
-
-    inline const std::string& getMaterialName() const
-    { return mMaterialName; }
-
-    inline Ogre::Real getRotationX() const
-    { return mRotationX; }
-
-    inline Ogre::Real getRotationY() const
-    { return mRotationY; }
-
-    inline Ogre::Real getRotationZ() const
-    { return mRotationZ; }
-
-private:
-    std::string mMeshName;
-    std::string mMaterialName;
-    Ogre::Real mRotationX;
-    Ogre::Real mRotationY;
-    Ogre::Real mRotationZ;
-};
-
-class TileSet
-{
-public:
-    TileSet(const Ogre::Vector3& scale);
-
-    std::vector<TileSetValue>& configureTileValues(TileVisual tileVisual);
-
-    const std::vector<TileSetValue>& getTileValues(TileVisual tileVisual) const;
-
-    inline const Ogre::Vector3& getScale() const
-    { return mScale; }
-
-private:
-    std::vector<std::vector<TileSetValue>> mTileValues;
-    Ogre::Vector3 mScale;
-};
 
 /*! \brief The class which stores the entire game state on the server and a subset of this on each client.
  *
